pyproject.toml
[project]
name = "tspng"
version = "0.0.4"
authors = [
  { name="Theia Scientific", email="support@theiascientific.com" },
]
description = "A Python package for manipulating PNG files with embedded JSON related to Machine Learning."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["Pillow"]

[project.urls]
"Homepage" = "https://github.com/Theia-Scientific/tspng"
"Bug Tracker" = "https://github.com/Theia-Scientific/tspng/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
tspng = "tspng.cli:app"

[project.optional-dependencies]
dev = ["build", "flake8", "pytest", "twine"]
cli = ["typer"]
